Theoretical PDF Generated!!!
I made a for loop that pulled a real number from the vector inside the InterdistRanges dictionary and plugs that into the nullpdf function instead of indexing into the vector inside the nullpdf function. I'm committing changes because now I'm changing to keys(interdist) for all the panels instead of a named vector containing the keys for this specific tumor.