Rewrote RunStatistics_Inter
It's not exactly the same, some of my checkpoints are not in here because I couldn't remember what they were. But the UnequalVarianceTTest is here. The hacky code for getting the pvalue and DoF is here. And it writes the comparisons to the CSV file. Next is to encode the MannWhitney Test and then do Gender and Chebyshev comparisons.