Updated RunStatistics_Inter Function
I added in all the comparisons for EGFR Norm, EGFR Alpha, EGFR beta, EGFR gamma, sarcomatoid status, rhabdoid status, pT (pairwise), and laterality. I also made grade pairwise comparisons. I had to update all the queries to be in sorted order for pT to get the groupby function to group in order from least invasive to most invasive, but the grade (which should be the only one affected by this) should still be in order. I will check the other 14 for loops for errors tomorrow.  And also recheck the grade for loops.